requirements.txt
pydantic

aiogram

python-dotenv

asyncpg

pillow